arm7 core pending rewrite
after long deliberation and performance monitoring i've concluded that
the core is too slow... fwiw, it's plenty fast now, but if we factor in
timers, maybe even sound down the line this will definitely be an
problem. i will re-do the decoding logic, and move the entire system to
a scheduler, better now than later.